+++ project/src/components/favorites/favorites.jsx
@@ -7,7 +7,7 @@ import FavoriteList from './favorite-list';
 import FavoritesEmpty from './favorites-empty';
 import offerProp from '../room/room.prop';
 import withLayout from '../hocs/with-layout';
-import Utils from '../../utils/utils';
+import Utils from '../../util/util';
 
 function Favorites(props) {
   const {places} = props;
