+++ project/src/store/places/places.js
@@ -1,6 +1,6 @@
 import {createReducer} from '@reduxjs/toolkit';
 import {SortByValues} from '../../const';
-import Utils from '../../utils/utils';
+import Utils from '../../util/util';
 
 import {
   changeCity,
